Complete a dataframe with useful variables for the calculation of disability-free life expectancies (DFLE)

From an input dataset containing some information on age, mortality rates and prevalences of disability (at least 'age', 'qx' or 'mx', and 'pix' must be within the input dataset), the function completes it with values of life expectancies and disability-free life expectancies. If some indispensable variables are missing (eg 'age'), the output is the same as the input dataset. If the input dataset contains mortality rates ('qx') but no prevalence of disabilities ('pix'), only total life-expectancy is calculed.

CompleteDFLEtable(tab)

Arguments

tab
a dataframe containing some of the variables

Value

a dataframe with all variables that can be calculated from the input dataframe

Details

Full headings for the names of variables in the output dataset can be found in the 'description_sullivan' dataframe (in the package).

If variables 'sex' and/or 'year' and/or 'categ' (an undefinite category variable that can be defined by the user) are in the input dataset, calculations are performed for each separate values of 'sex', 'year' and/or 'categ'.

Calculations follow example 1 in the sullivan manual (June 2007 version). See https://reves.site.ined.fr/en/resources/computation_online/sullivan/ for more explanations.

Examples

CompleteDFLEtable( sullivan[,c("year","age","mx","qx","pix")] )

#> # A tibble: 86 x 15
#>     year   age      mx      qx   pix agewidth     lx     Lx     Tx    ex   DFLx
#>    <dbl> <dbl>   <dbl>   <dbl> <dbl>    <dbl>  <dbl>  <dbl>  <dbl> <dbl>  <dbl>
#>  1  2004     0 3.69e-3 3.61e-3 0            1 1.00e5 99711. 8.14e6  81.4 99711.
#>  2  2004     1 3.83e-4 3.83e-4 0.048        1 9.96e4 99620. 8.04e6  80.7 94839.
#>  3  2004     2 1.98e-4 1.98e-4 0.048        1 9.96e4 99591. 7.94e6  79.7 94811.
#>  4  2004     3 1.43e-4 1.43e-4 0.048        1 9.96e4 99574. 7.84e6  78.8 94795.
#>  5  2004     4 2.15e-4 2.15e-4 0.048        1 9.96e4 99557. 7.74e6  77.8 94778.
#>  6  2004     5 1.42e-4 1.42e-4 0.03         1 9.95e4 99539. 7.64e6  76.8 96553.
#>  7  2004     6 1.04e-4 1.04e-4 0.03         1 9.95e4 99527. 7.54e6  75.8 96541.
#>  8  2004     7 1.89e-4 1.89e-4 0.03         1 9.95e4 99512. 7.44e6  74.8 96527.
#>  9  2004     8 8.52e-5 8.52e-5 0.03         1 9.95e4 99498. 7.34e6  73.8 96513.
#> 10  2004     9 1.01e-4 1.01e-4 0.03         1 9.95e4 99489. 7.25e6  72.8 96504.
#> # ... with 76 more rows, and 4 more variables: DFTx <dbl>, DFLEx <dbl>,
#> #   DLEx <dbl>, pctDFLEx <dbl>

CompleteDFLEtable( FRInseeMortalityForecast2016[FRInseeMortalityForecast2016$year %in% c(2013,2015,2020,2030),] )

#> # A tibble: 968 x 10
#>     year sex      age        qx agewidth        mx      lx     Lx       Tx    ex
#>    <dbl> <fct>  <dbl>     <dbl>    <dbl>     <dbl>   <dbl>  <dbl>    <dbl> <dbl>
#>  1  2013 female     0 0.00322          1 0.00322   100000  99743. 8501250.  85.0
#>  2  2013 female     1 0.000364         1 0.000364   99678. 99660. 8401507.  84.3
#>  3  2013 female     2 0.000168         1 0.000168   99642. 99634. 8301847.  83.3
#>  4  2013 female     3 0.000130         1 0.000130   99625. 99619. 8202214.  82.3
#>  5  2013 female     4 0.000127         1 0.000127   99612. 99606. 8102595.  81.3
#>  6  2013 female     5 0.000105         1 0.000105   99600. 99594. 8002989.  80.4
#>  7  2013 female     6 0.0000844        1 0.0000844  99589. 99585. 7903395.  79.4
#>  8  2013 female     7 0.0000820        1 0.0000820  99581. 99577. 7803810.  78.4
#>  9  2013 female     8 0.0000767        1 0.0000767  99573. 99569. 7704233.  77.4
#> 10  2013 female     9 0.0000656        1 0.0000656  99565. 99562. 7604664.  76.4
#> # ... with 958 more rows